+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Version 4.0
 - Simplified isort config
 - ``pyproject.toml`` and isolated builds adopted by default, issue #256
 - Added comment to ``setup.cfg`` template instructing about extra links, issue #268
+- Generate ``tox.ini`` by default, issue #296
 
 Version 3.3
 -----------